Lazily load libjawt if JNA getComponentId or getWindowId is used
These JNA calls will fail with an error complaining about a failure
to load libjawt.

It is only necessary to explicitly load libjawt if using one or other
of those JNA calls, and not for example in a JavaFX application.

Therefore, when invoking either of those methods, explicitly load
libjawt once and once only the first time it is needed.

This is implemented using the initialisation-on-demand holder idiom.